mgr: move handle_pyerror() from MgrPyModules to MgrPyModule
so we can reuse it in a more sane way.
